Move FrequencyPlot to /components/sharedPlots and add a check on which environment app is running to decide whether to use sample of condition in chosenClusters
