+++ src/pages/experiments/[experimentId]/data-processing/data-integration/components/fake_new_data.json
@@ -1,15 +1,1 @@
-[{"PC": 1, "percentVariance": 0.0695912662233713}, 
-{"PC": 2, "percentVariance": 0.0464235963048139}, 
-{"PC": 3, "percentVariance": 0.0409943294487813},
-{"PC": 4, "percentVariance": 0.0401269234549365},
-{"PC": 5, "percentVariance": 0.026034567971635}, 
-{"PC": 6, "percentVariance": 0.0222276879303816}, 
-{"PC": 7, "percentVariance": 0.0215141081950021}, 
-{"PC": 8, "percentVariance": 0.0203866847788484}, 
-{"PC": 9, "percentVariance": 0.0194361045684111}, 
-{"PC": 10, "percentVariance": 0.0188710842385607}, 
-{"PC": 11, "percentVariance": 0.0183440112280242}, 
-{"PC": 12, "percentVariance": 0.0182317232321607}, 
-{"PC": 13, "percentVariance": 0.0179837585001603}, 
-{"PC": 14, "percentVariance": 0.017882867022623}, 
-{"PC": 15, "percentVariance": 0.0177908259900759}, {"PC": 16, "percentVariance": 0.0176621482595031}, {"PC": 17, "percentVariance": 0.0176170586728534}, {"PC": 18, "percentVariance": 0.0174555880849617}, {"PC": 19, "percentVariance": 0.0174180432846133}, {"PC": 20, "percentVariance": 0.0173512013340646}, {"PC": 21, "percentVariance": 0.017260579355901}, {"PC": 22, "percentVariance": 0.0171797268225471}, {"PC": 23, "percentVariance": 0.017103435487742}, {"PC": 24, "percentVariance": 0.0170405432135146}, {"PC": 25, "percentVariance": 0.0169365391042983}, {"PC": 26, "percentVariance": 0.0169290719676287}, {"PC": 27, "percentVariance": 0.01687049739597}, {"PC": 28, "percentVariance": 0.0168331069010516}, {"PC": 29, "percentVariance": 0.0167849937299379}, {"PC": 30, "percentVariance": 0.0167707377515886}, {"PC": 31, "percentVariance": 0.0167222909834068}, {"PC": 32, "percentVariance": 0.0166360384033748}, {"PC": 33, "percentVariance": 0.0165978110924039}, {"PC": 34, "percentVariance": 0.0165756885803951}, {"PC": 35, "percentVariance": 0.0165443774624861}, {"PC": 36, "percentVariance": 0.0164812140708958}, {"PC": 37, "percentVariance": 0.0164573941892336}, {"PC": 38, "percentVariance": 0.0164319809163447}, {"PC": 39, "percentVariance": 0.0163884779595168}, {"PC": 40, "percentVariance": 0.0163647208511092}, {"PC": 41, "percentVariance": 0.0163081952112513}, {"PC": 42, "percentVariance": 0.016301594442556}, {"PC": 43, "percentVariance": 0.0162582683913705}, {"PC": 44, "percentVariance": 0.0161894672132257}, {"PC": 45, "percentVariance": 0.0161749043236022}, {"PC": 46, "percentVariance": 0.0161602173615978}, {"PC": 47, "percentVariance": 0.0161414499083575}, {"PC": 48, "percentVariance": 0.0161041148570463}, {"PC": 49, "percentVariance": 0.0160692972009101}, {"PC": 50, "percentVariance": 0.0160396861269541}]
+[{"PC": 1, "percentVariance": 0.0695912662233713}, {"PC": 2, "percentVariance": 0.0464235963048139}, {"PC": 3, "percentVariance": 0.0409943294487813},{"PC": 4, "percentVariance": 0.0401269234549365},{"PC": 5, "percentVariance": 0.026034567971635}, {"PC": 6, "percentVariance": 0.0222276879303816}, {"PC": 7, "percentVariance": 0.0215141081950021}, {"PC": 8, "percentVariance": 0.0203866847788484}, {"PC": 9, "percentVariance": 0.0194361045684111}, {"PC": 10, "percentVariance": 0.0188710842385607}, {"PC": 11, "percentVariance": 0.0183440112280242}, {"PC": 12, "percentVariance": 0.0182317232321607}, {"PC": 13, "percentVariance": 0.0179837585001603}, {"PC": 14, "percentVariance": 0.017882867022623}, {"PC": 15, "percentVariance": 0.0177908259900759}, {"PC": 16, "percentVariance": 0.0176621482595031}, {"PC": 17, "percentVariance": 0.0176170586728534}, {"PC": 18, "percentVariance": 0.0174555880849617}, {"PC": 19, "percentVariance": 0.0174180432846133}, {"PC": 20, "percentVariance": 0.0173512013340646}, {"PC": 21, "percentVariance": 0.017260579355901}, {"PC": 22, "percentVariance": 0.0171797268225471}, {"PC": 23, "percentVariance": 0.017103435487742}, {"PC": 24, "percentVariance": 0.0170405432135146}, {"PC": 25, "percentVariance": 0.0169365391042983}, {"PC": 26, "percentVariance": 0.0169290719676287}, {"PC": 27, "percentVariance": 0.01687049739597}, {"PC": 28, "percentVariance": 0.0168331069010516}, {"PC": 29, "percentVariance": 0.0167849937299379}, {"PC": 30, "percentVariance": 0.0167707377515886}, {"PC": 31, "percentVariance": 0.0167222909834068}, {"PC": 32, "percentVariance": 0.0166360384033748}, {"PC": 33, "percentVariance": 0.0165978110924039}, {"PC": 34, "percentVariance": 0.0165756885803951}, {"PC": 35, "percentVariance": 0.0165443774624861}, {"PC": 36, "percentVariance": 0.0164812140708958}, {"PC": 37, "percentVariance": 0.0164573941892336}, {"PC": 38, "percentVariance": 0.0164319809163447}, {"PC": 39, "percentVariance": 0.0163884779595168}, {"PC": 40, "percentVariance": 0.0163647208511092}, {"PC": 41, "percentVariance": 0.0163081952112513}, {"PC": 42, "percentVariance": 0.016301594442556}, {"PC": 43, "percentVariance": 0.0162582683913705}, {"PC": 44, "percentVariance": 0.0161894672132257}, {"PC": 45, "percentVariance": 0.0161749043236022}, {"PC": 46, "percentVariance": 0.0161602173615978}, {"PC": 47, "percentVariance": 0.0161414499083575}, {"PC": 48, "percentVariance": 0.0161041148570463}, {"PC": 49, "percentVariance": 0.0160692972009101}, {"PC": 50, "percentVariance": 0.0160396861269541}]
